comments + update

+++ src/libuc/int_fastdiv.hh
@@ -114,11 +114,11 @@ private:
 		}
 	}
 	friend int operator/(cint divident,const int_fastdiv& divisor) noex ;
-};
+} ;
 
 int operator/(cint n,const int_fastdiv& divisor) noex {
-	int q;
-	q = (((unsigned long long)((long long)divisor.M * (long long)n)) >> 32);
+	int	q ;
+	q = int(((unsigned long)((long)divisor.M * (long)n)) >> 32) ;
 	q += n * divisor.n_add_sign;
 	if (divisor.s >= 0) {
 	    /* we rely on this to be implemented as arithmetic shift */
